Flytter oppdater-DTO-ene og intern hjelpe-dto til services/moduler
