+++ src/@types/index.ts
@@ -29,10 +29,7 @@ export type { default as Organisasjon } from './organisasjon';
 export type { MedlPeriode } from './medlemskap';
 export type { default as Aktoer } from './aktoer';
 export type {
-  OppdaterAvgiftsgrunnlag,
-  AvgiftsgrunnlagInfo,
   Avgiftsperiode,
   Avgiftsgrunnlag,
-  OppdaterAvgiftsberegning,
   Avgiftsberegning,
 } from './trygdeavgift';
