+++ src/global.d.ts
@@ -34,10 +34,7 @@ declare module 'Domene' {
   export type Medlemskapsperiode = import('./@types').Medlemskapsperiode;
   export type OppdaterMedlemskapsperiode = import('./@types').OppdaterMedlemskapsperiode
   export type Familiemedlem = import('./@types').Familiemedlem;
-  export type OppdaterAvgiftsgrunnlag = import('./@types').OppdaterAvgiftsgrunnlag;
   export type Avgiftsgrunnlag = import('./@types').Avgiftsgrunnlag;
-  export type AvgiftsgrunnlagInfo = import('./@types').AvgiftsgrunnlagInfo;
-  export type OppdaterAvgiftsberegning = import('./@types').OppdaterAvgiftsberegning;
   export type Avgiftsberegning = import('./@types').Avgiftsberegning;
   export type Avgiftsperiode = import('./@types').Avgiftsperiode;
 }
